refactor: 도메인이 너무 좁은 파일 네이밍 개선 및 유틸함수, 상수 위치 변경(answerGenerator => baseballUtils)
