+++ docs/README.md
@@ -99,6 +99,7 @@
       => 음 뭔가 에러메시지 많아지면 도메인별로 나누는게 확장성 좋을듯. 다만, 지금은 에러메시지 많지 않으니 일단 유지하기.
 - [x] message
 - [x] type => 흠 뭔가 과도한 파일 분리같은데. 굳이 소프트코딩을 적용했어야 했나? 타입 분리가 엄청 크리티컬하진 않지만, 과도한 파일분리로 볼 수 있을듯. BaseballNumber 파일 내부 하단으로 옮기기!
+  - [x] 분리하다보니, 그냥 파일 이름을 잘못정해서 도메인이 애매하게 분리됨. answerGenerator => baseballNubmerUtils 변경
 
 ### Model
