+++ src/App.js
@@ -48,4 +48,6 @@ class App {
   }
 }
 
+new App().play();
+
 export default App;
