+++ src/constants/baseballGame.js
@@ -1,9 +1,3 @@
-export const BASEBALL_NUMBER = Object.freeze({
-  DIGIT: 3,
-  MAX: 9,
-  MIN: 1,
-});
-
 export const WINNING_CONDITION = Object.freeze({
   THREE_STRIKE: 3,
 });
