+++ src/constants/baseballNumber.js
@@ -1,0 +1,15 @@
+export const BASEBALL_NUMBER = Object.freeze({
+  DIGIT: 3,
+  MAX: 9,
+  MIN: 1,
+});
+
+export const TYPE = Object.freeze({
+  NUMBER: 'number',
+  STRING: 'string',
+});
+
+export const NUMBER = Object.freeze({
+  RADIX: 10,
+  ZERO: 0,
+});
